RazorPages is ambiguous - rename app MyRazorPages #4695 (#4775)
* Rename RazorPages to RazorPagesSample

* fix typo

* change dbname in startup

* revert change and slight rename
