+++ aspnetcore/mvc/razor-pages/index/sample/RazorPagesContacts/StartupAdvanced.cs
@@ -4,7 +4,7 @@ using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 
-namespace WebApplication2
+namespace RazorPagesContacts
 {
     public class Startup
     {
