+++ aspnetcore/mvc/razor-pages/index/sample/RazorPagesContacts/StartupDebug.cs
@@ -17,8 +17,8 @@ namespace RazorPagesContacts
 
         public void ConfigureServices(IServiceCollection services)
         {
-            services.AddDbContext<AppDbContext>(options => 
-                              options.UseInMemoryDatabase("name"));
+            services.AddDbContext<AppDbContext>(options =>
+                              options.UseInMemoryDatabase("InMemoryDb"));
             services.AddMvc();
         }
