+++ aspnetcore/mvc/razor-pages/index/sample/RazorPagesIntro/Pages/Index2.cshtml
@@ -1,5 +1,5 @@
 ﻿@page
-@using RazorPages
+@using RazorPagesIntro.Pages
 @model IndexModel2
 
 <h2>Separate page model</h2>
