+++ aspnetcore/tutorials/razor-pages/razor-pages-start/sample/RazorPagesMovie/Models/FileUpload.cs
@@ -1,7 +1,7 @@
 using Microsoft.AspNetCore.Http;
 using System.ComponentModel.DataAnnotations;
 
-namespace RazorPagesMovie.Models
+namespace RazorPagesSampleMovie.Models
 {
     public class FileUpload
     {
