+++ aspnetcore/tutorials/razor-pages/razor-pages-start/sample/RazorPagesMovie/Models/Schedule.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
 
-namespace RazorPagesMovie.Models
+namespace RazorPagesSampleMovie.Models
 {
     public class Schedule
     {
@@ -12,18 +12,18 @@ namespace RazorPagesMovie.Models
 
         public string PublicSchedule { get; set; }
 
-        [Display(Name="Public Schedule Size (bytes)")]
+        [Display(Name = "Public Schedule Size (bytes)")]
         [DisplayFormat(DataFormatString = "{0:N1}")]
         public long PublicScheduleSize { get; set; }
 
         public string PrivateSchedule { get; set; }
 
-        [Display(Name="Private Schedule Size (bytes)")]
+        [Display(Name = "Private Schedule Size (bytes)")]
         [DisplayFormat(DataFormatString = "{0:N1}")]
         public long PrivateScheduleSize { get; set; }
 
-        [Display(Name="Uploaded (UTC)")]
+        [Display(Name = "Uploaded (UTC)")]
         [DisplayFormat(DataFormatString = "{0:F}")]
-        public DateTime UploadDT { get; set;}
+        public DateTime UploadDT { get; set; }
     }
 }
