+++ aspnetcore/tutorials/razor-pages/razor-pages-start/sample/RazorPagesMovie/Pages/Contact.cshtml.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 
-namespace RazorPagesMovie.Pages
+namespace RazorPagesSampleMovie.Pages
 {
     public class ContactModel : PageModel
     {
