+++ aspnetcore/tutorials/razor-pages/razor-pages-start/sample/RazorPagesMovie/Pages/_ViewImports.cshtml
@@ -1,3 +1,3 @@
 @using RazorPagesMovie
-@namespace RazorPagesMovie.Pages
+@namespace RazorPagesSampleMovie.Pages
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
