+++ aspnetcore/tutorials/razor-pages/razor-pages-start/sample/RazorPagesMovie/Program.cs
@@ -6,7 +6,7 @@ using Microsoft.Extensions.Logging;
 using RazorPagesMovie.Models;
 using System;
 
-namespace RazorPagesMovie
+namespace RazorPagesSampleMovie
 {
     public class Program
     {
