+++ aspnetcore/tutorials/razor-pages/razor-pages-start/snapshot_sample/RazorPagesMovie/Models/MovieNoEF.cs
@@ -4,7 +4,7 @@
 #region snippet_MovieNoEF
 using System;
 
-namespace RazorPagesMovie.Models
+namespace RazorPagesSampleMovie.Models
 {
     public class Movie
     {
