+++ aspnetcore/tutorials/razor-pages/razor-pages-start/snapshot_sample/RazorPagesMovie/Pages/Schedules/Index.cshtml.cs
@@ -7,7 +7,7 @@ using Microsoft.EntityFrameworkCore;
 using RazorPagesMovie.Models;
 using RazorPagesMovie.Utilities;
 
-namespace RazorPagesMovie.Pages.Schedules
+namespace RazorPagesSampleMovie.Pages.Schedules
 {
     public class IndexModel : PageModel
     {
